feat: Implement barebones addresspack
This implementation does not allow for any customisation via the CLI.

The default options for SQLite are: Synchronisation normal and journal
mode WAL.

If the implementation finds a database with a `trailer` table it will
consume files according to that. So, it effectively tries to “resume” a
loading process that was interrupted. If the database is corrupted due
to a bad interruption or tampering with the wal journal, the behaviour
is undefined.

Finally, the data is expected to be decompressed and inside a `data`
directory.

Signed-off-by: Arnau Siches <asiches@gmail.com>